third_party/WebKit/LayoutTests/http/tests/websocket/tests/hybi/handshake-error-expected.txt

CONSOLE ERROR: line 20: WebSocket connection to 'ws://127.0.0.1:8880/websocket/tests/hybi/handshake-error' failed: Error during WebSocket handshake: No response code found in status line: ThisWillCauseHandshakeError
Handshake error test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Closed.
PASS ws.readyState is 3
PASS successfullyParsed is true

TEST COMPLETE
